sql大幅度改进 减少数据库连接（长到死的SQL）

+++ src/main/java/com/superman/superman/service/TaoBaoApiService.java
@@ -24,5 +24,5 @@ public interface TaoBaoApiService {
 
      * @return
      */
-    Integer  countWaitTb(List list);
+    Long countWaitTb(List list);
 }
